Inject logger and Provider Container in middleware
This is necessary for accessing the Riverpod providers through the Provider Container in the route handlers, as well as logging requests.

This patch also includes the start of the API routes, which currently has a simple Health Check response and a basic sample of Dart Frog middleware for creating an auth guard.

<!-- ps-id: d79d27d9-3173-430b-88d4-c5769e5ab35a -->