Add logger initialization to main entry point
This facilitates logging within the app if the environment is not production.

This adds a new entry point to the dart frog server via `main.dart`. This allows us to initialize the Fimber logger and print the App Config values for debugging purposes. This can be useful for understanding if environment variables are incorrect or missing upon starting the server.

<!-- ps-id: 277dff93-427d-4ee7-96bc-6f92781a06fc -->