Refactor the Service Provider Error as subclass
This facilitates explicit error handling and code reuse by extending the generic `LittleHelpBookError`.

This removes the code generation from this error in favor of a more simple and straightforward error that is defined by the superclass. This allows for flexibility in updating the message and status code, which can be sent in the final response of the route if desired.

<!-- ps-id: 4fd19663-c8aa-4440-9256-83e79b5c4e7d -->