Expose API Provider route
This facilitates the retrieval of service providers from the datastore for the consumer.

This exposes the route for retrieving a list of service providers. Ideally, this should be a paginated response, but the underlying Airtable client library does not seem to support pagination. Based on the Airtable documents, the maximum retrieval is 100 records, which means this endpoint does not currently fetch all records.

<!-- ps-id: 438cd42e-bcb6-45dc-940d-a4d76efc4985 -->