LHB-13 Add auth guard middleware to API routes
This ensures that all incoming requests are authorized to access the API routes.

This api key will be shared with the frontend app and provided in the Authorization header of requests. This patch adds the new api key to the App Config and updates the middleware to check for this value.

<!-- ps-id: 4517b294-43ff-45ae-8107-96846438ccc3 -->